requirements.txt
aiogram>=3.7
yt-dlp
python-dotenv
pyrogram
qrcode
